+++ src/kairos/authentication/authentication-composition.ts
@@ -1,4 +1,6 @@
 import { securityConfig } from '@/config/security';
+
+import { usersCollection } from '@/framework/mongodb/collection/users-collection';
 import { createLogin, type GenerateAccessToken } from '@/kairos/authentication/application/login';
 import { type FindOneByEmail } from '@/kairos/authentication/domain/user-credentials-repository';
 import { createFindOneByEmail } from '@/kairos/authentication/infrastructure/repository/find-one-by-email';
@@ -11,9 +13,7 @@ import { passwordHasher } from '@/kairos/authentication/infrastructure/security/
 import { createVerifyPassword } from '@/kairos/authentication/infrastructure/security/password/verify-password';
 import { loginRequestConstraints } from '@/kairos/authentication/interface/http/login-request';
 import { systemClock } from '@/kairos/shared/clock';
-import { validateRequest } from '@/kairos/shared/http/http-composition';
-
-import { usersCollection } from '@/framework/mongodb/collection/users-collection';
+import { validateRequest } from '@/kairos/shared/http/request-validation';
 import { createSecretKey, randomUUID } from 'node:crypto';
 
 const signingKey = createSecretKey(Buffer.from(securityConfig.appKey.replace('base64:', ''), 'base64'));
