+++ src/kairos/shared/http/map-result-to-http.ts
@@ -5,7 +5,6 @@ import {
 } from '@/kairos/shared/http/result-to-http';
 import type { Result } from '@/kairos/shared/kernel/result';
 import { createNormalizer, type Normalizer } from '@koala-ts/framework';
-import { builtInConstraints, createValidationMiddleware, createValidator } from '@koala-ts/framework/validator';
 
 export const normalize: Normalizer = createNormalizer([]);
 
@@ -16,7 +15,3 @@ export function mapResultToHttp<T, E extends { type: string }>(
   const resultToHttpMapper = createResultToHttpMapper({ normalize });
   return resultToHttpMapper(result, httpMapping);
 }
-
-const validate = createValidator({ constraints: builtInConstraints });
-
-export const validateRequest = createValidationMiddleware({ validate });
