+++ src/kairos/shared/http/request-validation.ts
@@ -1,0 +1,4 @@
+import { builtInConstraints, createValidationMiddleware, createValidator } from '@koala-ts/framework/validator';
+
+const validate = createValidator({ constraints: builtInConstraints });
+export const validateRequest = createValidationMiddleware({ validate });
